+++ lib/data/user_32/user_32_api.dart
@@ -146,35 +146,35 @@ class User32Api {
   void simulateCharacter({
     required int char,
     required bool isCapsLockActive,
-    required KeyboardKeyEvent keyEvent,
+    required KeyboardEventType eventType,
   }) {
     KeyData keyData = getKeyData(char);
     if (isCapsLockActive &&
         keyData.requirements.singleOrNull == KeyboardEvent.shift) {
       simulateKeyboardEvent(
         keyboardEvent: keyData.keyboardEvent,
-        keyEvent: keyEvent,
+        eventType: eventType,
       );
       return;
     }
-    if (keyEvent == KeyboardKeyEvent.down) {
+    if (eventType == KeyboardEventType.down) {
       for (var key in keyData.requirements) {
         simulateKeyboardEvent(
           keyboardEvent: key,
-          keyEvent: KeyboardKeyEvent.down,
+          eventType: KeyboardEventType.down,
           repeatOnKeyDown: false,
         );
       }
     }
     simulateKeyboardEvent(
       keyboardEvent: keyData.keyboardEvent,
-      keyEvent: keyEvent,
+      eventType: eventType,
     );
-    if (keyEvent == KeyboardKeyEvent.up) {
+    if (eventType == KeyboardEventType.up) {
       for (var key in keyData.requirements) {
         simulateKeyboardEvent(
           keyboardEvent: key,
-          keyEvent: KeyboardKeyEvent.up,
+          eventType: KeyboardEventType.up,
           repeatOnKeyDown: false,
         );
       }
@@ -183,16 +183,28 @@ class User32Api {
 
   void simulateKeyboardEvent({
     required KeyboardEvent keyboardEvent,
-    required KeyboardKeyEvent keyEvent,
+    required KeyboardEventType eventType,
     bool repeatOnKeyDown = true,
   }) {
-    _performKeyPress(keyboardEvent, keyEvent);
+    _repeatingKeyboardEvent(
+      keyPress: () => _performKeyPress(keyboardEvent, eventType),
+      eventType: eventType,
+      repeatOnKeyDown: repeatOnKeyDown,
+    );
+  }
+
+  void _repeatingKeyboardEvent({
+    required void Function() keyPress,
+    required KeyboardEventType eventType,
+    bool repeatOnKeyDown = true,
+  }) {
+    keyPress();
     if (repeatOnKeyDown) {
-      switch (keyEvent) {
-        case KeyboardKeyEvent.down:
-          _startRepeat(() => _performKeyPress(keyboardEvent, keyEvent));
+      switch (eventType) {
+        case KeyboardEventType.down:
+          _startRepeat(keyPress);
           break;
-        case KeyboardKeyEvent.up:
+        case KeyboardEventType.up:
           _stopRepeat();
           break;
       }
@@ -201,7 +213,7 @@ class User32Api {
 
   void _performKeyPress(
     KeyboardEvent keyboardEvent,
-    KeyboardKeyEvent keyEvent,
+    KeyboardEventType eventType,
   ) {
     int inputCount = 1;
     Pointer<INPUT> inputs = calloc<INPUT>(inputCount);
@@ -210,7 +222,7 @@ class User32Api {
       inputs.ref.u.ki
         ..wVk = keyboardEvent.value
         ..wScan = 0
-        ..dwFlags = keyEvent.value
+        ..dwFlags = eventType.value
         ..time = 0
         ..dwExtraInfo = nullptr;
       int inputSize = sizeOf<INPUT>();