+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+# v1.3.0
+
+- Added custom key mapping configuration file 📋
+- Added many new button actions and key combinations 🔑
+- Added support for triggers 🎮
+
 # v1.0.0
 
 - Initial release ✅